Random US States and Capitals Quiz Generator
Generates 35 unique versions of a quiz that tests the user's knowledge of the capitals of each of the 50 US states. Each version has a randomized order of questions and this program is designed for a teacher to provide to their class of 35 students to prevent cheating in an exam setting (since the order of questions in each version of the test is randomized and unique).